refactor uint64 code
