+++ src/api.ts
@@ -60,13 +60,15 @@ export interface Vector<T> extends Iterable<T> {
   append(value: T): Vector<T>
 }
 
-export type PaddedPieceSize = New<{ PaddedPieceSize: bigint }>
+export type uint64 = bigint
+
+export type PaddedPieceSize = New<{ PaddedPieceSize: uint64 }>
 
 /**
  * `UnpaddedPieceSize` is the size of a piece, in bytes.
  * @see https://github.com/filecoin-project/go-state-types/blob/ff2ed169ff566458f2acd8b135d62e8ca27e7d0c/abi/piece.go#L10C4-L11
  */
-export type UnpaddedPieceSize = New<{ UnpaddedPieceSize: bigint }>
+export type UnpaddedPieceSize = New<{ UnpaddedPieceSize: uint64 }>
 
 export type Fr23Padded = New<{ Fr23Padded: Uint8Array }>
