+++ src/index.js
@@ -1,7 +1,7 @@
 import * as API from './api.js'
 import * as Node from './node.js'
 import * as Segment from './segment.js'
-import { log2Ceil, log2Floor } from './math.js'
+import { log2Ceil } from './uint64.js'
 
 export const { Uint64Size, ChecksumSize } = Segment
 
@@ -11,7 +11,9 @@ export const { Uint64Size, ChecksumSize } = Segment
  * {@link Uint64Size} for the segment `offset`, {@link Uint64Size} for the
  * segment `size` and {@link Segment.ChecksumSize} for the segment `checksum`.
  */
-export const EntrySize = Node.Size + Uint64Size + Uint64Size + ChecksumSize
+export const EntrySize = BigInt(
+  Node.Size + Uint64Size + Uint64Size + ChecksumSize
+)
 
 /**
  * Function that returns the maximum number of index entries in a given deal
@@ -27,13 +29,13 @@ export const EntrySize = Node.Size + Uint64Size + Uint64Size + ChecksumSize
  * the size of an entry (in bytes), rounded up to the nearest power of 2. The minimum
  * return value is 4.
  *
- * @param {bigint|number} size - The size of the deal in bytes.
+ * @param {API.uint64} size - The size of the deal in bytes.
  * @returns {number} - The maximum number of index entries for a given deal size.
  */
 export const maxIndexEntriesInDeal = (size) => {
   // The raw result is the size of the deal divided by 2048 times the size of an
   // entry, rounded up to the nearest power of 2.
-  const n = Math.pow(2, log2Ceil(BigInt(size) / 2048n / BigInt(EntrySize)))
+  const n = Math.pow(2, log2Ceil(size / 2048n / EntrySize))
 
   // If the number is less than 4, return 4, otherwise return actual number.
   return n < 4 ? 4 : n
